refactor: pool unstake/remove
