+++ packages/web/src/views/pool/pool-remove/components/remove-liquidity-select-list/remove-liquidity-select-list-item/RemoveLiquiditySelectListItem.styles.ts
@@ -1,0 +1,116 @@
+import { fonts } from "@constants/font.constant";
+import mixins from "@styles/mixins";
+import { inputStyle } from "@views/pool/pool-remove/components/remove-liquidity/RemoveLiquidity.styles";
+
+import styled from "@emotion/styled";
+import { media } from "@styles/media";
+
+export const RemoveLiquiditySelectListItemWrapper = styled.li<{
+  selected: boolean;
+}>`
+  ${({ theme }) => inputStyle(theme)};
+  ${mixins.flexbox("row", "center", "flex-start")};
+  width: 100%;
+  gap: 8px;
+  background-color: ${({ theme }) => theme.color.background20};
+  border: 1px solid
+    ${({ theme, selected }) =>
+      selected ? theme.color.border15 : theme.color.border02};
+  border-radius: 8px;
+  padding: 15px;
+  ${fonts.body12};
+  color: ${({ theme }) => theme.color.text03};
+  transition: all 0.3s ease;
+
+  ${media.mobile} {
+    padding: 11px;
+  }
+  .logo-wrapper {
+    ${mixins.flexbox("row", "center", "flex-start")};
+    gap: 5px;
+  }
+  input[type="checkbox"] + label:before {
+    background-color: ${({ theme }) => theme.color.background12};
+  }
+
+  .left-content {
+    ${mixins.flexbox("row", "center", "flex-start")};
+    gap: 5px;
+  }
+  .liquidity-value {
+    margin-left: auto;
+    color: ${({ theme }) => theme.color.text02};
+  }
+  .liquidity-value-fake {
+    position: absolute;
+    visibility: hidden;
+    opacity: 0;
+    pointer-events: none;
+    right: 0;
+  }
+
+  .hover-info {
+    &,
+    & * {
+      width: 16px;
+      height: 16px;
+    }
+    cursor: pointer;
+    .icon-info {
+      * {
+        fill: ${({ theme }) => theme.color.icon03};
+      }
+    }
+  }
+  label {
+    margin-right: 3px;
+  }
+  .token-id {
+    cursor: default;
+  }
+`;
+
+export const TooltipWrapperContent = styled.div`
+  width: 268px;
+  ${mixins.flexbox("column", "flex-start", "flex-start")};
+  gap: 8px;
+  ${fonts.body12}
+  > div {
+    &:not(:first-of-type) {
+      padding: 4px 0;
+    }
+    .title {
+      color: ${({ theme }) => theme.color.text04};
+    }
+    .value {
+      ${mixins.flexbox("row", "center", "center")};
+      gap: 8px;
+      color: ${({ theme }) => theme.color.text02};
+      img {
+        width: 20px;
+        height: 20px;
+      }
+    }
+  }
+  .divider {
+    width: 100%;
+    border-top: 1px solid ${({ theme }) => theme.color.border01};
+    padding: 0 !important;
+  }
+  .unstake-description {
+    color: ${({ theme }) => theme.color.text04};
+    ${fonts.p4}
+    padding: 0 !important;
+    text-align: left;
+    align-items: start;
+  }
+`;
+
+export const TokenValueWrapper = styled.div`
+  ${mixins.flexbox("row", "center", "space-between")};
+  width: 100%;
+`;
+export const TokenTitleWrapper = styled.div`
+  ${mixins.flexbox("row", "center", "space-between")};
+  width: 100%;
+`;
