+++ packages/web/src/views/pool/pool-remove/components/remove-liquidity-select-list/remove-liquidity-select-list-item/RemoveLiquiditySelectListItem.tsx
