+++ packages/web/src/views/pool/pool-remove/components/remove-position-modal/RemovePositionModal.styles.ts
