+++ packages/web/src/views/pool/pool-unstake/PoolUnstake.tsx
@@ -4,7 +4,6 @@ import { useTranslation } from "react-i18next";
 import Footer from "@components/common/footer/Footer";
 import BreadcrumbsContainer from "@containers/breadcrumbs-container/BreadcrumbsContainer";
 import HeaderContainer from "@containers/header-container/HeaderContainer";
-import UnstakeLiquidityContainer from "@containers/unstake-position-container/UnstakePositionContainer";
 import useRouter from "@hooks/common/use-custom-router";
 import { useLoading } from "@hooks/common/use-loading";
 import { useWindowSize } from "@hooks/common/use-window-size";
@@ -12,6 +11,7 @@ import { useGnotToGnot } from "@hooks/token/use-gnot-wugnot";
 import { DeviceSize } from "@styles/media";
 import { useGetPoolDetailByPath } from "src/react-query/pools";
 
+import UnstakeLiquidityContainer from "./containers/unstake-position-container/UnstakePositionContainer";
 import UnstakeLiquidityLayout from "./UnstakeLiquidityLayout";
 
 const PoolUnstake: React.FC = () => {
