+++ packages/web/src/views/pool/pool-unstake/components/select-unstake-result/SelectUnstakeResult.stories.tsx
@@ -1,0 +1,15 @@
+import { ComponentMeta, ComponentStory } from "@storybook/react";
+
+import SelectUnstakeResult from "./SelectUnstakeResult";
+
+export default {
+  title: "pool/pool-unstake/SelectUnstakeResult",
+  component: SelectUnstakeResult,
+} as ComponentMeta<typeof SelectUnstakeResult>;
+
+const Template: ComponentStory<typeof SelectUnstakeResult> = args => {
+  return <SelectUnstakeResult {...args} />;
+};
+
+export const Default = Template.bind({});
+Default.args = {};
