+++ packages/web/src/views/pool/pool-unstake/components/unstake-liquidity/UnstakeLiquidity.stories.tsx
@@ -1,0 +1,17 @@
+import { ComponentMeta, ComponentStory } from "@storybook/react";
+
+import UnstakeLiquidity from "./UnstakeLiquidity";
+
+export default {
+  title: "pool/pool-unstake/UnstakeLiquidity",
+  component: UnstakeLiquidity,
+} as ComponentMeta<typeof UnstakeLiquidity>;
+
+const Template: ComponentStory<typeof UnstakeLiquidity> = args => {
+  return (
+    <UnstakeLiquidity {...args} />
+  );
+};
+
+export const Default = Template.bind({});
+Default.args = {};
